% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bank}
\alias{bank}
\title{Bank Data}
\format{
A data frame with 100 observations on 2 variables
\describe{
\item{arrival_time}{Interval Arrival time for each customer}
\item{service_time}{Service time for each customer}
}
}
\usage{
data(bank)
}
\description{
This dataset simulates customer arrival and service times in a bank.
The bank data frame has 100 variables
on 2 variables
}
\examples{
## For examples see example(Multiserver)
}
\keyword{datasets}
