% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multiserver.R
\name{Multiserver}
\alias{Multiserver}
\title{Simulate a first come, first serve queuing system}
\usage{
Multiserver(Arrivals, ServiceTimes, NumServers = 1)
}
\arguments{
\item{Arrivals}{numeric; positive for the arrival time}

\item{ServiceTimes}{numeric; positive for the service time}

\item{NumServers}{a positive integer for the number of servers in the queuing system}
}
\value{
A tibble:tibble() with information about the queuing system
}
\description{
\code{Multiserver} can simulate customer queuing times on a first-come-first-serve basis by determining their arrival and service times.
}
\examples{
data(bank)
with(bank, Multiserver(arrival_time, service_time, 5))
}
