% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bank}
\alias{bank}
\title{Bank Data}
\format{
A data frame with 100 customers and 2 variables
\describe{
\item{arrival_time}{Customer Arrival time}
\item{service_time}{Service time for each customer}
}
}
\usage{
data(bank)
}
\description{
This dataset simulates customer arrival and service times in a bank.
The bank data frame measures the arrival and service times of 100 customers
}
\examples{
## see example(Multiserver)
}
\keyword{datasets}
