% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multiserver.R
\name{Multiserver}
\alias{Multiserver}
\title{Simulate a first come, first serve queuing system}
\usage{
Multiserver(Arrivals, ServiceTimes, NumServers = 1)
}
\arguments{
\item{Arrivals}{numeric; strictly positive for inter-arrival times}

\item{ServiceTimes}{numeric; strictly positive for service times}

\item{NumServers}{positive integer; number of server we have in the queue system}
}
\value{
A tibble:tibble() with information about the queuing system
}
\description{
\code{Multiserver} is used to simulate a first come, first serve multi-server queuing system by sorting the inter-arrival and service times.
}
\examples{
data(bank)
with(bank, Multiserver(arrival_time, service_time, 2))
}
